Split CohostEndpointTestBase in half, with "devenv" stuff only at the VS layer
